src/projects/findxai/pyproject.toml
[project]
name = "findxai"
version = "0.1.0"
description = ""
authors = [
    {name = "Arthur",email = "dnthanh1298@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli] (>=1.6.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "protobuf (>=6.30.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
